Update LinuxGuide.md
Removed fvwm from a list item for virtual machines. FVWM stands for F? Virtual Window Manager and its an X.Org/X11 window manager, has nothing to do with virtual machines. 

The other change moved picom from the Wayland Compositors session to the more general Linux System session, particularly because the letter has a Window Manager index.
The reason for it is that picom is a compositor for X.Org/X11, not for Wayland. That word has different meanings when it comes to those display protocols. On the former it refers to a software to be (optionally) used alongside a window manager to add some visual effects to it, like window animations, blur, rounded corners and so on. On the latter it refers to the window manger manager itself, which may or may not have said visual effects built directly into it.